+++ packages/appstore/src/components/appstore-banner/appstore-banner.scss
@@ -10,7 +10,7 @@
 
         @include desktop {
             display: grid;
-            grid-template-columns: 1fr 1fr;
+            grid-template-columns: 0.9fr 1.1fr;
         }
 
         @include mobile {
@@ -56,7 +56,7 @@
 
             @include mobile {
                 &__text {
-                    width: 81%;
+                    width: 90%;
                 }
             }
         }
