+++ packages/shared/src/utils/cfd/cfd.ts
@@ -1,6 +1,7 @@
 import { DetailsOfEachMT5Loginid, GetAccountStatus, LandingCompany } from '@deriv/api-types';
 import { localize } from '@deriv/translations';
 import { CFD_PLATFORMS } from '../platform';
+import { Jurisdiction, JURISDICTION_MARKET_TYPES } from '../constants/jurisdictions-config';
 
 let CFD_text_translated: { [key: string]: () => void };
 
@@ -465,3 +466,48 @@ export const getAuthenticationStatusInfo = (account_status: GetAccountStatus): T
 
 export const mt5_community_url =
     'https://community.deriv.com/t/log-in-using-mt5-pc-or-mobile-app-application-guideline/49622';
+
+export const getFormattedJurisdictionCode = (jurisdiction_code: string) => {
+    let formatted_label = '';
+
+    switch (jurisdiction_code) {
+        case Jurisdiction.SVG:
+            formatted_label = localize('SVG');
+            break;
+        case Jurisdiction.BVI:
+            formatted_label = localize('BVI');
+            break;
+        case Jurisdiction.LABUAN:
+            formatted_label = localize('Labuan');
+            break;
+        case Jurisdiction.VANUATU:
+            formatted_label = localize('Vanuatu');
+            break;
+        case Jurisdiction.MALTA_INVEST:
+            formatted_label = localize('Malta');
+            break;
+        default:
+            formatted_label = jurisdiction_code?.toUpperCase();
+            break;
+    }
+
+    return formatted_label;
+};
+
+export const getFormattedJurisdictionMarketTypes = (jurisdiction_market_type: string) => {
+    let formatted_market_type = '';
+
+    switch (jurisdiction_market_type) {
+        case 'synthetic': // need to remove this once we have the correct market type from BE
+        case JURISDICTION_MARKET_TYPES.DERIVED:
+            formatted_market_type = localize('Derived');
+            break;
+        case JURISDICTION_MARKET_TYPES.FINANCIAL:
+            formatted_market_type = localize('Financial');
+            break;
+        default:
+            formatted_market_type = jurisdiction_market_type?.toUpperCase();
+            break;
+    }
+    return formatted_market_type;
+};
