+++ packages/shared/src/utils/constants/jurisdictions-config.ts
@@ -1,5 +1,3 @@
-import { localize } from '@deriv/translations';
-
 export const Jurisdiction = Object.freeze({
     SVG: 'svg',
     BVI: 'bvi',
@@ -13,33 +11,6 @@ export const Platforms = Object.freeze({
     MT5: 'mt5',
     DERIVEZ: 'derivez',
 });
-
-export const getFormattedJurisdictionCode = (jurisdiction_code: string) => {
-    let formatted_label = '';
-
-    switch (jurisdiction_code) {
-        case Jurisdiction.SVG:
-            formatted_label = localize('SVG');
-            break;
-        case Jurisdiction.BVI:
-            formatted_label = localize('BVI');
-            break;
-        case Jurisdiction.LABUAN:
-            formatted_label = localize('Labuan');
-            break;
-        case Jurisdiction.VANUATU:
-            formatted_label = localize('Vanuatu');
-            break;
-        case Jurisdiction.MALTA_INVEST:
-            formatted_label = localize('Malta');
-            break;
-        default:
-            formatted_label = jurisdiction_code?.toUpperCase();
-            break;
-    }
-
-    return formatted_label;
-};
 
 export const DBVI_COMPANY_NAMES = {
     bvi: { name: 'Deriv (BVI) Ltd', tnc_url: 'tnc/deriv-(bvi)-ltd.pdf' },
@@ -55,21 +26,3 @@ export const JURISDICTION_MARKET_TYPES = {
     FINANCIAL: 'financial',
     DERIVED: 'derived',
 } as const;
-
-export const getFormattedJurisdictionMarketTypes = (jurisdiction_market_type: string) => {
-    let formatted_market_type = '';
-
-    switch (jurisdiction_market_type) {
-        case 'synthetic': // need to remove this once we have the correct market type from BE
-        case JURISDICTION_MARKET_TYPES.DERIVED:
-            formatted_market_type = localize('Derived');
-            break;
-        case JURISDICTION_MARKET_TYPES.FINANCIAL:
-            formatted_market_type = localize('Financial');
-            break;
-        default:
-            formatted_market_type = jurisdiction_market_type?.toUpperCase();
-            break;
-    }
-    return formatted_market_type;
-};
